requirements.txt
torch>=2.1.0
torchvision
torchaudio
transformers>=4.44.2
accelerate>=0.33.0
datasets>=2.20.0
huggingface_hub>=0.24.6
tokenizers>=0.19.1
safetensors>=0.4.5
sentencepiece>=0.2.0
vllm>=0.10.0
einops>=0.8.0
tqdm>=4.66.5
fire>=0.7.0
python-dotenv>=1.0.0
jsonlines>=4.0.0
numpy>=1.26.4
pandas>=2.3.0
